feat(camera): Add PSF interface and Airy disk model - Introduce PSF interface and AiryDisk implementation for point spread function modeling - Refactor Aperture and CircularAperture to be templated and support PSF creation - Update CameraModel to manage PSF, allow custom or aperture-based PSF, and use float for focal length/fstop - Update SensorModel to provide pixel pitch and sensor size accessors - Update handle and example usage for new float-based API

+++ include/huira/handles/camera_handle.hpp
@@ -21,8 +21,8 @@ namespace huira {
         CameraModelHandle() = delete;
         using Handle<CameraModel<TSpectral>>::Handle;
 
-        void set_focal_length(double focal_length) const;
-        void set_fstop(double fstop) const;
+        void set_focal_length(float focal_length) const;
+        void set_fstop(float fstop) const;
 
         FrameBuffer<TSpectral> make_frame_buffer() const;
